refactor(compositions): intégrer useDiscordMessage dans page.tsx
- Créer hook useDiscordMessage pour gérer toute la logique Discord
- Extraire les états Discord (sendingDiscord, discordSentStatus, customMessages, etc.)
- Extraire les fonctions Discord (sendMessage, formatMatchInfo, insertMention, saveCustomMessage)
- Extraire les useEffect pour charger membres/canaux et vérifier statut
- Réduire significativement la taille du fichier page.tsx
