+++ src/app/compositions/page.tsx
@@ -47,7 +47,6 @@ import { FirestorePlayerService } from "@/lib/services/firestore-player-service"
 import { CompositionService } from "@/lib/services/composition-service";
 import { CompositionDefaultsService } from "@/lib/services/composition-defaults-service";
 import { Player } from "@/types/team-management";
-import { Match } from "@/types";
 import { AuthGuard } from "@/components/AuthGuard";
 import { USER_ROLES } from "@/lib/auth/roles";
 import {
@@ -80,6 +79,7 @@ import { useCurrentPhase } from "@/hooks/useCurrentPhase";
 import { AvailablePlayerItem } from "@/components/compositions/AvailablePlayerItem";
 import { useCompositionState } from "@/hooks/useCompositionState";
 import { CompositionDialogs } from "@/components/compositions/CompositionDialogs";
+import { useDiscordMessage } from "@/hooks/useDiscordMessage";
 
 // TabPanel remplacé par CompositionTabPanel
 
@@ -296,7 +296,7 @@ export default function CompositionsPage() {
     useState(false);
   const [isResetting, setIsResetting] = useState(false);
   const [isApplyingDefaults, setIsApplyingDefaults] = useState(false);
-  
+
   // draggedPlayerId et dragOverTeamId sont maintenant gérés par useCompositionDragDrop
   // État pour la recherche de joueurs
   const [searchQuery, setSearchQuery] = useState<string>("");
@@ -321,34 +321,6 @@ export default function CompositionsPage() {
   const [locations, setLocations] = useState<
     Array<{ id: string; name: string }>
   >([]);
-  const [sendingDiscord, setSendingDiscord] = useState<Record<string, boolean>>(
-    {}
-  );
-  const [discordSentStatus, setDiscordSentStatus] = useState<
-    Record<string, { sent: boolean; sentAt?: string; customMessage?: string }>
-  >({});
-  const [customMessages, setCustomMessages] = useState<Record<string, string>>(
-    {}
-  );
-  const [confirmResendDialog, setConfirmResendDialog] = useState<{
-    open: boolean;
-    teamId: string | null;
-    matchInfo: string | null;
-    channelId?: string;
-  }>({ open: false, teamId: null, matchInfo: null });
-  const [discordMembers, setDiscordMembers] = useState<
-    Array<{ id: string; username: string; displayName: string }>
-  >([]);
-  const [discordChannels, setDiscordChannels] = useState<
-    Array<{ id: string; name: string }>
-  >([]);
-  const [mentionAnchor, setMentionAnchor] = useState<{
-    teamId: string;
-    anchorEl: HTMLElement;
-    startPos: number;
-  } | null>(null);
-  const [mentionQuery, setMentionQuery] = useState<string>("");
-  const saveTimeoutRef = React.useRef<Record<string, NodeJS.Timeout>>({});
 
   // getDiscordStatus est maintenant importé depuis @/lib/compositions/discord-utils
 
@@ -638,84 +610,7 @@ export default function CompositionsPage() {
     void loadLocations();
   }, []);
 
-  // Charger les membres Discord pour l'autocomplete
-  useEffect(() => {
-    const loadDiscordMembers = async () => {
-      try {
-        console.log("[Compositions] Chargement des membres Discord...");
-        const response = await fetch("/api/discord/members", {
-          method: "GET",
-          credentials: "include",
-        });
-        console.log("[Compositions] Réponse status:", response.status);
-        if (response.ok) {
-          const result = await response.json();
-          console.log("[Compositions] Résultat:", result);
-          if (result.success) {
-            console.log(
-              "[Compositions] Membres reçus:",
-              result.members?.length || 0
-            );
-            setDiscordMembers(result.members || []);
-          } else {
-            console.error(
-              "[Compositions] Erreur dans la réponse:",
-              result.error
-            );
-          }
-        } else {
-          const errorText = await response.text();
-          console.error(
-            "[Compositions] Erreur HTTP:",
-            response.status,
-            errorText
-          );
-        }
-      } catch (error) {
-        console.error(
-          "[Compositions] Erreur lors du chargement des membres Discord:",
-          error
-        );
-      } finally {
-      }
-    };
-    void loadDiscordMembers();
-  }, []);
-
-  // Charger les canaux Discord
-  useEffect(() => {
-    const loadDiscordChannels = async () => {
-      try {
-        const response = await fetch("/api/discord/channels", {
-          method: "GET",
-          credentials: "include",
-          headers: {
-            "Content-Type": "application/json",
-          },
-        });
-        if (response.ok) {
-          const result = await response.json();
-          if (result.success) {
-            setDiscordChannels(result.channels || []);
-          } else {
-            console.error(
-              "Erreur lors du chargement des canaux Discord:",
-              result.error
-            );
-          }
-        } else {
-          const errorData = await response.json();
-          console.error(
-            "Erreur HTTP lors du chargement des canaux Discord:",
-            errorData
-          );
-        }
-      } catch (error) {
-        console.error("Erreur lors du chargement des canaux Discord:", error);
-      }
-    };
-    void loadDiscordChannels();
-  }, []);
+  // Le chargement des membres et canaux Discord est maintenant géré par useDiscordMessage
 
   const [availabilities, setAvailabilities] = useState<{
     masculin?: Record<string, { available?: boolean; comment?: string }>;
@@ -951,68 +846,36 @@ export default function CompositionsPage() {
   // Grouper les équipes par type (masculin/féminin)
   const equipesByType = useEquipesByType(filteredEquipes);
 
-  // Vérifier le statut d'envoi des messages Discord
-  useEffect(() => {
-    const checkDiscordStatus = async () => {
-      if (!selectedJournee || !selectedPhase) return;
-
-      const allTeams = [...equipesByType.masculin, ...equipesByType.feminin];
-      if (allTeams.length === 0) return;
-
-      // Un seul appel API pour toutes les équipes
-      const teamIds = allTeams.map((equipe) => equipe.team.id).join(",");
-      try {
-        const response = await fetch(
-          `/api/discord/check-message-sent?teamIds=${encodeURIComponent(
-            teamIds
-          )}&journee=${selectedJournee}&phase=${encodeURIComponent(
-            selectedPhase
-          )}`,
-          {
-            method: "GET",
-            credentials: "include",
-          }
-        );
-        if (response.ok) {
-          const result = await response.json();
-          if (result.success && result.results) {
-            const statusMap: Record<
-              string,
-              { sent: boolean; sentAt?: string; customMessage?: string }
-            > = {};
-            const customMessagesMap: Record<string, string> = {};
-
-            Object.entries(result.results).forEach(([teamId, status]) => {
-              const statusData = status as {
-                sent: boolean;
-                sentAt?: string;
-                customMessage?: string;
-              };
-              statusMap[teamId] = statusData;
-              if (statusData.customMessage) {
-                customMessagesMap[teamId] = statusData.customMessage;
-              }
-            });
-
-            setDiscordSentStatus(statusMap);
-            setCustomMessages((prev) => ({ ...prev, ...customMessagesMap }));
-          }
-        }
-      } catch (error) {
-        console.error(
-          "Erreur lors de la vérification du statut Discord:",
-          error
-        );
-      }
-    };
-
-    void checkDiscordStatus();
-  }, [
+  // Utiliser le hook useDiscordMessage pour gérer toute la logique Discord
+  const discordMessage = useDiscordMessage({
     selectedJournee,
     selectedPhase,
-    equipesByType.masculin,
-    equipesByType.feminin,
-  ]);
+    equipesByType,
+    locations,
+    cleanTeamName,
+    formatDivision,
+  });
+
+  // Extraire les valeurs du hook
+  const {
+    sendingDiscord,
+    discordSentStatus,
+    customMessages,
+    setCustomMessages,
+    confirmResendDialog,
+    setConfirmResendDialog,
+    discordMembers,
+    discordChannels,
+    mentionAnchor,
+    setMentionAnchor,
+    mentionQuery,
+    setMentionQuery,
+    sendMessage: handleSendDiscordMessage,
+    formatMatchInfo,
+    insertMention,
+    saveCustomMessage: handleSaveCustomMessage,
+    saveTimeoutRef,
+  } = discordMessage;
 
   const hasAssignedPlayers = useMemo(
     () =>
@@ -1124,273 +987,7 @@ export default function CompositionsPage() {
     [selectedJournee, selectedPhase]
   );
 
-  // Fonction pour sauvegarder le message personnalisé
-  const handleSaveCustomMessage = useCallback(
-    async (
-      teamId: string,
-      customMessage: string,
-      journee: number | null,
-      phase: string | null
-    ) => {
-      if (!journee || !phase) return;
-
-      try {
-        const response = await fetch("/api/discord/update-custom-message", {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          credentials: "include",
-          body: JSON.stringify({
-            teamId,
-            journee,
-            phase,
-            customMessage,
-          }),
-        });
-        if (response.ok) {
-          const result = await response.json();
-          if (result.success) {
-            // Mettre à jour l'état local
-            setCustomMessages((prev) => ({ ...prev, [teamId]: customMessage }));
-            setDiscordSentStatus((prev) => ({
-              ...prev,
-              [teamId]: { ...prev[teamId], customMessage },
-            }));
-          }
-        }
-      } catch (error) {
-        console.error(
-          "Erreur lors de la sauvegarde du message personnalisé:",
-          error
-        );
-      }
-    },
-    []
-  );
-
-  // Fonction pour insérer une mention dans le message
-  const insertMention = useCallback(
-    (
-      teamId: string,
-      startPos: number,
-      member: { id: string; displayName: string; username: string }
-    ) => {
-      setCustomMessages((prev) => {
-        const currentMessage = prev[teamId] || "";
-        const textBeforeAt = currentMessage.substring(0, startPos);
-        const textAfterAt = currentMessage.substring(startPos);
-        // Trouver où se termine la partie à remplacer (jusqu'au prochain espace, saut de ligne ou fin)
-        const match = textAfterAt.match(/^@[^\s\n]*/);
-        const endPos = startPos + (match ? match[0].length : 1);
-        const textAfterMention = currentMessage.substring(endPos);
-
-        const mention = `<@${member.id}>`;
-        const newMessage = textBeforeAt + mention + textAfterMention;
-
-        // Sauvegarder immédiatement après l'insertion
-        if (selectedJournee && selectedPhase) {
-          handleSaveCustomMessage(
-            teamId,
-            newMessage,
-            selectedJournee,
-            selectedPhase
-          );
-        }
-
-        return { ...prev, [teamId]: newMessage };
-      });
-      setMentionAnchor(null);
-    },
-    [selectedJournee, selectedPhase, handleSaveCustomMessage]
-  );
-
-  // Fonction pour formater le message d'informations du match
-  const formatMatchInfo = useCallback(
-    (
-      match: Match | null,
-      teamPlayers: Player[],
-      teamLocationId?: string,
-      teamName?: string,
-      isFemale?: boolean,
-      epreuve?: EpreuveType | null
-    ) => {
-      if (!match) return null;
-
-      const dayNames = [
-        "Dimanche",
-        "Lundi",
-        "Mardi",
-        "Mercredi",
-        "Jeudi",
-        "Vendredi",
-        "Samedi",
-      ];
-      const matchDate =
-        match.date instanceof Date ? match.date : new Date(match.date);
-      const dayName = dayNames[matchDate.getDay()];
-
-      // Nettoyer le nom de l'équipe et ajouter "F" pour les équipes féminines
-      let cleanName = teamName ? cleanTeamName(teamName) : "";
-      if (isFemale && cleanName && !cleanName.endsWith("F")) {
-        cleanName = `${cleanName}F`;
-      }
-      const division = formatDivision(match.division || "");
-      const opponent = match.opponent || match.opponentClub || "";
-
-      // Utiliser le lieu de l'équipe (résoudre l'ID en nom depuis la collection locations)
-      // Ne jamais utiliser match.location car il peut contenir des données incorrectes
-      // On n'affiche rien si le lieu n'est pas trouvé dans la liste des locations
-      let location = "";
-      if (teamLocationId) {
-        const teamLocation = locations.find((l) => l.id === teamLocationId);
-        if (teamLocation) {
-          location = teamLocation.name;
-        }
-        // Si le lieu n'est pas trouvé, on laisse location vide (on n'affiche rien)
-      }
-
-      const isHome = match.isHome ? "Domicile" : "Extérieur";
-
-      // Pour le championnat de Paris, grouper les joueurs et ajouter le numéro de groupe
-      // Utiliser isParisChampionship depuis le hook au lieu de vérifier epreuve directement
-      const isParisMatch = epreuve === "championnat_paris";
-      let playersList: string;
-
-      if (isParisMatch) {
-        // Trier les joueurs par points décroissants
-        const sortedPlayers = [...teamPlayers].sort((a, b) => {
-          const pointsA = a.points ?? 0;
-          const pointsB = b.points ?? 0;
-          return pointsB - pointsA; // Décroissant
-        });
-
-        // Grouper par 3 et ajouter le numéro de groupe
-        const playersWithGroups: string[] = [];
-        for (let i = 0; i < sortedPlayers.length; i += 3) {
-          const groupNumber = Math.floor(i / 3) + 1;
-          const group = sortedPlayers.slice(i, i + 3);
-          group.forEach((player) => {
-            playersWithGroups.push(
-              `🔹 ${player.firstName} ${player.name} (Groupe ${groupNumber})`
-            );
-          });
-        }
-        playersList = playersWithGroups.join("\n");
-      } else {
-        // Pour les autres championnats, affichage normal
-        playersList = teamPlayers
-          .map((p) => `🔹 ${p.firstName} ${p.name}`)
-          .join("\n");
-      }
-
-      // Collecter toutes les mentions Discord de tous les joueurs
-      // Filtrer uniquement les mentions qui correspondent à des membres existants sur le serveur
-      const allDiscordMentions: string[] = [];
-      const validMemberIds = new Set(discordMembers.map((m) => m.id));
-      teamPlayers.forEach((player) => {
-        if (player.discordMentions && player.discordMentions.length > 0) {
-          // Ajouter uniquement les mentions valides (format <@USER_ID>)
-          player.discordMentions.forEach((mentionId) => {
-            // Vérifier que l'ID existe encore dans la liste des membres Discord
-            if (validMemberIds.has(mentionId)) {
-              allDiscordMentions.push(`<@${mentionId}>`);
-            }
-          });
-        }
-      });
-
-      // Construire le message avec les mentions Discord si elles existent
-      const parts = [
-        `${cleanName} – ${division} – ${opponent}`,
-        location,
-        `${dayName} – ${isHome}`,
-        playersList,
-      ];
-
-      // Ajouter les mentions Discord après la liste des joueurs
-      if (allDiscordMentions.length > 0) {
-        parts.push(allDiscordMentions.join(" "));
-      }
-
-      return parts.filter(Boolean).join("\n");
-    },
-    [locations, discordMembers]
-  );
-
-  // Fonction pour envoyer le message Discord
-  const handleSendDiscordMessage = useCallback(
-    async (
-      teamId: string,
-      content: string,
-      journee: number | null,
-      phase: string | null,
-      channelId?: string
-    ) => {
-      if (!journee || !phase) return;
-
-      setSendingDiscord((prev) => ({ ...prev, [teamId]: true }));
-      try {
-        const customMessage = customMessages[teamId] || "";
-        const response = await fetch("/api/discord/send-message", {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          credentials: "include",
-          body: JSON.stringify({
-            content,
-            teamId,
-            journee,
-            phase,
-            customMessage,
-            channelId,
-          }),
-        });
-        if (response.ok) {
-          const result = await response.json();
-          if (result.success) {
-            // Mettre à jour le statut local
-            setDiscordSentStatus((prev) => {
-              const existing = prev[teamId];
-              const newStatus: {
-                sent: boolean;
-                sentAt: string;
-                customMessage?: string;
-              } = {
-                sent: true,
-                sentAt: new Date().toISOString(),
-                ...(customMessage || existing?.customMessage
-                  ? {
-                      customMessage:
-                        customMessage || existing?.customMessage || "",
-                    }
-                  : {}),
-              };
-              return {
-                ...prev,
-                [teamId]: newStatus,
-              };
-            });
-          } else {
-            const error = await response.json();
-            alert(
-              `Erreur lors de l'envoi: ${error.error || "Erreur inconnue"}`
-            );
-          }
-        } else {
-          const error = await response.json();
-          alert(`Erreur lors de l'envoi: ${error.error || "Erreur inconnue"}`);
-        }
-      } catch (error) {
-        console.error("Erreur lors de l'envoi du message Discord:", error);
-        alert("Erreur lors de l'envoi du message Discord");
-      } finally {
-        setSendingDiscord((prev) => ({ ...prev, [teamId]: false }));
-      }
-    },
-    [customMessages]
-  );
+  // Les fonctions Discord sont maintenant gérées par useDiscordMessage
 
   const compositionSummary = useMemo(() => {
     // Pour le championnat de Paris, utiliser toutes les équipes (masculin + féminin)
@@ -1495,7 +1092,6 @@ export default function CompositionsPage() {
     setDefaultCompositions,
     canDropPlayer,
   });
-
 
   const handleResetButtonClick = useCallback(() => {
     if (!canResetButton) {
@@ -3082,7 +2678,6 @@ export default function CompositionsPage() {
             </Box>
           </>
         ) : null}
-
 
         {(!selectedJournee || !selectedPhase) && (
           <Card>
